Add payment methods, B2B portal, AI Intelligence modules; fix 11 bugs from QA audit
- AppContext: default currency CAD, repairHistory/aiMetrics persistence, getEnabledModules guard, redeemCoupon case-insensitive, coupon management (addCoupon/updateCoupon/deleteCoupon)
- Sidebar: AI Intelligence section (AI Guardian, Workflow Automation, AI Assistant), Account & Admin section (Subscription, Payment Methods, B2B Portal, Settings); hide empty sections; optional chaining for currentUser
- Login: remove demo credentials exposure
- Dashboard: monthly revenue stat (not all-time), remove unused totalRevenue
- Inventory: parseFloat for tax rate (QST 9.975% precision)
- POS: tax calculated on discounted price (CRA compliance fix); barcode null guard
- SalesOrders: dynamic currency symbol via getCurrencySymbol(); remove all hardcoded CA$
- WorkflowAutomation: new AI Intelligence add-on module with trigger/action rule builder
- AIAssistant: new AI chat module with live data queries
- PaymentMethods: new billing page with Stripe/PayPal/Square/Wave/Wise/Interac/Cheque connect UI and invoice history
- B2BPortalSettings: distributor/wholesaler portal configuration with shareable link
- B2BOrderPortal: public-facing dark wholesale catalogue portal that submits to Sales Orders
- App.js: routes for /workflow-automation, /ai-assistant, /payment-methods, /b2b-portal, /b2b-order (public)
- Header: page titles for all new routes

+++ arwa-inventory-pro/src/components/Inventory/Inventory.js
@@ -46,7 +46,7 @@ function ProductModal({ product, suppliers, onClose, onSave }) {
       salePrice: parseFloat(form.salePrice) || 0,
       stock: parseInt(form.stock) || 0,
       minStock: parseInt(form.minStock) || 0,
-      tax: parseInt(form.tax) || 0,
+      tax: parseFloat(form.tax) || 0,
     });
   };
